+++ mysql/datadog_checks/mysql/statement_samples.py
@@ -33,8 +33,8 @@ from datadog_checks.base.utils.tracking import tracked_method
 
 from .util import (
     DatabaseConfigurationError,
+    ManagedAuthConnectionMixin,
     StatementTruncationState,
-    connect_with_session_variables,
     get_truncation_state,
     warning_with_tags,
 )
@@ -188,12 +188,12 @@ ExplainState = namedtuple('ExplainState', ['strategy', 'error_code', 'error_mess
 EMPTY_EXPLAIN_STATE = ExplainState(strategy=None, error_code=None, error_message=None)
 
 
-class MySQLStatementSamples(DBMAsyncJob):
+class MySQLStatementSamples(ManagedAuthConnectionMixin, DBMAsyncJob):
     """
     Collects statement samples and execution plans.
     """
 
-    def __init__(self, check, config, connection_args):
+    def __init__(self, check, config, connection_args_provider, uses_managed_auth=False):
         collection_interval = float(config.statement_metrics_config.get('collection_interval', 1))
         if collection_interval <= 0:
             collection_interval = 1
@@ -210,7 +210,9 @@ class MySQLStatementSamples(DBMAsyncJob):
         )
         self._config = config
         self._version_processed = False
-        self._connection_args = connection_args
+        self._connection_args_provider = connection_args_provider
+        self._uses_managed_auth = uses_managed_auth
+        self._db_created_at = 0
         self._last_check_run = 0
         self._db = None
         self._check = check
@@ -272,16 +274,6 @@ class MySQLStatementSamples(DBMAsyncJob):
             else:
                 self._global_status_table = "performance_schema.global_status"
             self._version_processed = True
-
-    def _get_db_connection(self):
-        """
-        lazy reconnect db
-        pymysql connections are not thread safe so we can't reuse the same connection from the main check
-        :return:
-        """
-        if not self._db:
-            self._db = connect_with_session_variables(**self._connection_args)
-        return self._db
 
     def _close_db_conn(self):
         if self._db: