+++ mysql/tests/test_unit.py
@@ -292,7 +292,6 @@ def test_replication_check_status(
                 'database_hostname:stubbed.hostname',
                 'database_instance:stubbed.hostname',
                 'dd.internal.resource:database_instance:stubbed.hostname',
-                'ddagenthostname:stubbed.hostname',
             },
         ),
         (
@@ -304,7 +303,6 @@ def test_replication_check_status(
                 'database_hostname:stubbed.hostname',
                 'database_instance:stubbed.hostname',
                 'dd.internal.resource:database_instance:stubbed.hostname',
-                'ddagenthostname:stubbed.hostname',
             },
         ),
         (
@@ -315,7 +313,6 @@ def test_replication_check_status(
                 'database_hostname:stubbed.hostname',
                 'database_instance:stubbed.hostname',
                 'dd.internal.resource:database_instance:stubbed.hostname',
-                'ddagenthostname:stubbed.hostname',
             },
         ),
         (
@@ -327,7 +324,6 @@ def test_replication_check_status(
                 'database_hostname:stubbed.hostname',
                 'database_instance:stubbed.hostname',
                 'dd.internal.resource:database_instance:stubbed.hostname',
-                'ddagenthostname:stubbed.hostname',
             },
         ),
     ],