+++ postgres/datadog_checks/postgres/health.py
@@ -49,10 +49,11 @@ class PostgresHealth(Health):
         :param kwargs: Additional keyword arguments to include in the event.
         """
         super().submit_health_event(
-            name,
-            status,
+            name=name,
+            status=status,
             # If we have an error parsing the config we may not have tags yet
-            self.check.tags if hasattr(self.check, 'tags') else [],
-            database_identifier=self.check.database_identifier,
+            tags=self.check.tags if hasattr(self.check, 'tags') else [],
+            database_instance=self.check.database_identifier,
+            ddagenthostname=self.check.agent_hostname,
             **kwargs,
         )