+++ sqlserver/tests/test_database_metrics.py
@@ -101,9 +101,8 @@ def test_sqlserver_file_stats_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             db, state, logical_name, file_location, *metric_values = result
@@ -188,9 +187,8 @@ def test_sqlserver_ao_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_ao_availability_groups:
             (
@@ -310,9 +308,8 @@ def test_sqlserver_availability_groups_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             ag, availability_group_name, synchronization_health_desc, *metric_values = result
@@ -407,9 +404,8 @@ def test_sqlserver_database_replication_stats_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             ag, availability_group_name, replica_server_name, synchronization_state_desc, *metric_values = result
@@ -551,9 +547,8 @@ def test_sqlserver_availability_replicas_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             (
@@ -622,9 +617,8 @@ def test_sqlserver_fci_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             node_name, status, failover_cluster, *metric_values = result
@@ -676,9 +670,8 @@ def test_sqlserver_primary_log_shipping_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             primary_id, primary_server, primary_db, *metric_values = result
@@ -742,9 +735,8 @@ def test_sqlserver_secondary_log_shipping_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             secondary_server, secondary_db, secondary_id, primary_server, primary_db, *metric_values = result
@@ -795,9 +787,8 @@ def test_sqlserver_server_state_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             metrics = zip(server_state_metrics.metric_names()[0], result)
@@ -845,9 +836,8 @@ def test_sqlserver_tempdb_file_space_usage_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             database_id, *metric_values = result
@@ -933,9 +923,8 @@ def test_sqlserver_index_usage_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             for row in result:
@@ -1061,9 +1050,8 @@ def test_sqlserver_db_fragmentation_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             for row in result:
@@ -1150,9 +1138,8 @@ def test_sqlserver_os_schedulers_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             scheduler_id, parent_node_id, *metric_values = result
@@ -1221,9 +1208,8 @@ def test_sqlserver_os_tasks_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             scheduler_id, *metric_values = result
@@ -1293,9 +1279,8 @@ def test_sqlserver_master_files_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             db, database, file_id, file_type, file_location, database_files_state_desc, size, state = result
@@ -1375,9 +1360,8 @@ def test_sqlserver_database_files_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for db, result in zip(AUTODISCOVERY_DBS, mocked_results):
             for row in result:
@@ -1444,9 +1428,8 @@ def test_sqlserver_table_size_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for mock_row in mocked_results:
             table_name, schema, database, *metric_values = mock_row
@@ -1477,11 +1460,11 @@ def test_sqlserver_database_stats_metrics(
     }
 
     mocked_results = [
-        ('master', 'master', 'ONLINE', 'SIMPLE', 0, False, False, False),
-        ('tempdb', 'tempdb', 'ONLINE', 'SIMPLE', 0, False, False, False),
-        ('model', 'model', 'ONLINE', 'FULL', 0, False, False, False),
-        ('msdb', 'msdb', 'ONLINE', 'SIMPLE', 0, False, False, False),
-        ('datadog_test', 'datadog_test', 'ONLINE', 'FULL', 0, False, False, False),
+        ('master', 'master', 'ONLINE', 'SIMPLE', 0, 0, False, False, False),
+        ('tempdb', 'tempdb', 'ONLINE', 'SIMPLE', 0, 0, False, False, False),
+        ('model', 'model', 'ONLINE', 'FULL', 0, 0, False, False, False),
+        ('msdb', 'msdb', 'ONLINE', 'SIMPLE', 0, 0, False, False, False),
+        ('datadog_test', 'datadog_test', 'ONLINE', 'FULL', 0, 0, False, False, False),
     ]
 
     sqlserver_check = SQLServer(CHECK_NAME, init_config, [instance_docker_metrics])
@@ -1506,12 +1489,17 @@ def test_sqlserver_database_stats_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
-            db, database, database_state_desc, database_recovery_model_desc, *metric_values = result
+            (
+                db,
+                database,
+                database_state_desc,
+                database_recovery_model_desc,
+                *metric_values,
+            ) = result
             metrics = zip(database_stats_metrics.metric_names()[0], metric_values)
             expected_tags = [
                 f'db:{db}',
@@ -1576,9 +1564,8 @@ def test_sqlserver_database_backup_metrics(
         tags = sqlserver_check._config.tags + [
             "database_hostname:{}".format("stubbed.hostname"),
             "database_instance:{}".format("stubbed.hostname"),
-            "ddagenthostname:{}".format("stubbed.hostname"),
             "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+            "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
         ]
         for result in mocked_results:
             db, database, *metric_values = result
@@ -1621,9 +1608,8 @@ def test_sqlserver_xe_session_metrics(
     expected_tags = sqlserver_check._config.tags + [
         "database_hostname:{}".format("stubbed.hostname"),
         "database_instance:{}".format("stubbed.hostname"),
-        "ddagenthostname:{}".format("stubbed.hostname"),
         "dd.internal.resource:database_instance:{}".format("stubbed.hostname"),
-        "sqlserver_servername:{}".format(sqlserver_check.static_info_cache.get(STATIC_INFO_SERVERNAME)),
+        "sqlserver_servername:{}".format(sqlserver_check.static_info_cache[STATIC_INFO_SERVERNAME].lower()),
     ]
     expected_tags.append('session_name:datadog')
     aggregator.assert_metric("sqlserver.xe.session_status", value=1, tags=expected_tags)