+++ checkpoint_harmony_endpoint/CHANGELOG.md
@@ -2,3 +2,9 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.0 / 2025-10-02
+
+***Added***:
+
+* Initial Release ([#21149](https://github.com/DataDog/integrations-core/pull/21149))
+* Bump datadog-checks-base to 37.21.0 ([#21477](https://github.com/DataDog/integrations-core/pull/21477))