+++ silverstripe_cms/CHANGELOG.md
@@ -2,7 +2,14 @@
 
 <!-- towncrier release notes start -->
 
-## 1.4.1 / 2025-09-24
+## 1.5.0 / 2025-10-02
+
+***Added***:
+
+* Bump Python to 3.13 ([#21161](https://github.com/DataDog/integrations-core/pull/21161))
+* Bump datadog-checks-base to 37.21.0 ([#21477](https://github.com/DataDog/integrations-core/pull/21477))
+
+## 1.4.1 / 2025-09-24 / Agent 7.71.0
 
 ***Fixed***: